Fix rainbow restart
This fixes rainbow application on save.

At some point in the past callbacks probably changed and now
`on_after_commit` is called before uci commit for some reason (wtf). The
correct handler is now `on_after_apply` looking in to other LuCI
applications but that is called only if `apply_on_parse` is set to
`true` or `false`. In default `apply_on_parse` is set to `nil` so we
have to set it.