Omnia LED names changed in upstream kernel
Updating model for Omnia to reflect those changes and to make Rainbow
work again.

Make sure we have correct names of the LEDs and let's add new colors
available in rainbow while we are at it. There is one issue though. UCI
can't handle sections containing dash (-), so I had to replace them with
underscore (_). We need to make appropriate adjustments in Rainbow as
well.